refactor(community): migrate imports and remove
