+++ src/app/providers/RoleProvider.tsx
@@ -1,7 +1,7 @@
 import { createContext, useContext, useMemo, type PropsWithChildren } from "react"
 import { useParams } from "@tanstack/react-router"
-import { useUserRolesInServer, useServerById } from "@/shared/queries/community/community.queries"
-import type { Role } from "@/shared/queries/community/community.types"
+import { useUserRolesInServer, useServerById } from "@/shared/queries/community"
+import type { Role } from "@/shared/queries/community"
 import { aggregateRolePermissions } from "@/shared/lib/permissions"
 import { Permissions } from "@/shared/models/permissions"
 import { useAuth } from "@/app/providers/KeycloakAuthProvider"
