+++ src/pages/server-settings/feature/PageCreateRoleFeature.tsx
@@ -3,13 +3,13 @@ import { zodResolver } from "@hookform/resolvers/zod"
 import { addRoleFormSchema } from "@/shared/zod/add-role"
 import type z from "zod"
 import { AddRoleForm } from "@/shared/forms/AddRole"
-import { useCreateRole } from "@/shared/queries/community/community.queries"
+import { useCreateRole } from "@/shared/queries/community"
 import { useNavigate } from "@tanstack/react-router"
 import { useEffect } from "react"
 import { toast } from "sonner"
 import { useTranslation } from "react-i18next"
 import { useQueryClient } from "@tanstack/react-query"
-import { communityKeys } from "@/shared/queries/community/community.queries"
+import { communityKeys } from "@/shared/queries/community"
 import { PERMISSION_METADATA } from "@/shared/lib/permissions"
 
 interface PageCreateRoleFeatureProps {
