+++ src/routes/servers/$id/route.tsx
@@ -3,7 +3,7 @@ import { useEffect, useState, useMemo, useCallback, useRef } from "react"
 import { useSidebarContent } from "@/app/providers/SidebarContentProvider"
 import { ServerProfile, ServerProfileSkeleton } from "@/shared/components/ServerProfile"
 import ServerChannels, { ServerChannelsSkeleton } from "@/shared/components/ServerChannels"
-import { useServerById, useServerMembers } from "@/shared/queries/community/community.queries"
+import { useServerById, useServerMembers } from "@/shared/queries/community"
 import { useUsersBatch } from "@/shared/queries/user/user.queries"
 import TopBarServers from "@/shared/components/TopBarServers"
 import MembersSidebar from "@/shared/components/MembersSidebar"
